product_review_analyzer/backend/requirements.txt
pyramid
sqlalchemy
psycopg2-binary
requests
google-generativeai
wsgicors
waitress